requirements.txt
django<5
djangorestframework>=3.0
pytz
py-dateutil
